Multithreading: code refactoring

+++ multithreading/deadlock/src/main/java/com/bsgfb/cdp/deadlock/model/BlockingRunner.java
@@ -50,7 +50,7 @@ public class BlockingRunner implements Runnable {
 
     private void work() {
         try {
-            TimeUnit.MILLISECONDS.sleep((long) 1000);
+            TimeUnit.MILLISECONDS.sleep(1000L);
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
